refactor(gallery): create abstract class for db
create abstract class for db

MODULE3-PART2-dynamodb
