+++ backend/api/backend/interfaces/user.ts
@@ -1,5 +1,8 @@
-export interface DynamoUser {
+import { ImagesArray } from './image';
+
+export interface IUser {
   email: string;
   password: string;
   salt: string;
+  images?: ImagesArray;
 }
