+++ backend/api/backend/signup/handler.ts
@@ -4,20 +4,18 @@ import { log } from '@helper/logger';
 import crypto from 'crypto';
 import { SignupManager } from './signup.manager';
 import { APIGatewayProxyHandlerV2 } from 'aws-lambda';
-import { DbService } from '../services/db-service';
 import { AuthService } from '../services/auth';
+import { DynamoDB } from '../services/dynamo.service';
 
 const secret = process.env.SECRET;
+const dbService = new DynamoDB();
+const authService = new AuthService();
 
 export const signup: APIGatewayProxyHandlerV2 = async (event) => {
   try {
     const manager = new SignupManager();
-
     const { email, password } = JSON.parse(event.body!);
     const salt = crypto.randomBytes(16).toString('hex');
-
-    const dbService = new DbService();
-    const authService = new AuthService();
 
     await manager.user.createUser(email, password, salt, dbService);
     log(`User ${email} is created`);
