+++ backend/api/backend/signup/signup.user.ts
@@ -1,9 +1,7 @@
-import { DbService } from '../services/db-service';
-
 export class SignupUser {
-  async createUser(email: string, password: string, salt: string, dbService: DbService): Promise<void> {
+  async createUser(email: string, password: string, salt: string, dbService: any): Promise<void> {
     try {
-      await dbService.user.createDynamoUser(email, password, salt);
+      await dbService.createUser(email, password, salt);
     } catch (e) {
       throw { errorMessage: 'Could not create a new user in database.', statusCode: 400 };
     }
