+++ backend/api/backend/upload/handler.ts
@@ -6,11 +6,12 @@ import { IFileData } from '../interfaces/file';
 import jwt from 'jsonwebtoken';
 import { UploadManager } from './upload.manager';
 import { FileService } from '../services/file.service';
+import { DynamoDB } from '../services/dynamo.service';
 
 const secret = process.env.SECRET;
 const s3ImageDirectory = process.env.S3_IMAGE_DIRECTORY;
 const fileService = new FileService();
-const dynamoTable = 'module3_part2';
+const dbService = new DynamoDB();
 
 export const upload: APIGatewayProxyHandlerV2 = async (event) => {
   try {
@@ -23,7 +24,7 @@ export const upload: APIGatewayProxyHandlerV2 = async (event) => {
     const user = userEmail.split('@')[0];
 
     const imageMetadata = manager.getMetadata(fileService, data, type);
-    const imageArray = await manager.getImageArray(userEmail, dynamoTable);
+    const imageArray = await manager.getImagesArray(userEmail, dbService);
 
     imageArray.push({
       filename: filename,
@@ -32,7 +33,7 @@ export const upload: APIGatewayProxyHandlerV2 = async (event) => {
       date: new Date(),
     });
 
-    await manager.updateDynamoUser(userEmail, dynamoTable, imageArray);
+    await manager.updateUserInDB(userEmail, imageArray, dbService);
     await manager.uploadImageToS3(data, `${user}/${filename}`, s3ImageDirectory!);
     return createResponse(200);
   } catch (e) {
