+++ backend/api/defaults/add-defaults.db.ts
@@ -1,6 +1,10 @@
-import { DbService } from './db-service';
+import { DynamoDB } from '../backend/services/dynamo.service';
+import { defaultUsersArray } from './default.users';
 
 const imagesDir = './api/backend/images';
-const dbService = new DbService();
+const dbService = new DynamoDB();
 
-(async () => await dbService.startDb(imagesDir!))();
+(async () => {
+  await dbService.addDefaultUsersToDB(defaultUsersArray);
+  await dbService.addImagesDataToDB(imagesDir);
+})();
